chore: continuing to perfect test helper
